move file to correct directory
